# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciamento de transações e estatísticas.",
        "title": "Desafio Itau Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/estatistica": {
            "get": {
                "description": "Obtém estatísticas das transações dos últimos 60 segundos",
                "produces": [
                    "application/json"
                ],
                "summary": "Retorna estatísticas das transações",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Intervalo de busca em segundos",
                        "name": "intervaloBusca",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.StatisticsResponse"
                        }
                    }
                }
            }
        },
        "/transacao": {
            "post": {
                "description": "Registra uma transação, validando os critérios necessários",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Cria uma nova transação",
                "parameters": [
                    {
                        "description": "Dados da transação",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.TransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "description": "Apaga todas as transações armazenadas",
                "summary": "Remove todas as transações",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "dtos.StatisticsResponse": {
            "type": "object",
            "properties": {
                "avg": {
                    "type": "number"
                },
                "count": {
                    "type": "integer"
                },
                "max": {
                    "type": "number"
                },
                "min": {
                    "type": "number"
                },
                "sum": {
                    "type": "number"
                }
            }
        },
        "dtos.TransactionRequest": {
            "type": "object",
            "properties": {
                "dataHora": {
                    "type": "string"
                },
                "valor": {
                    "type": "number"
                }
            }
        }
    }
}